+++ apps/backoffice-v2/src/common/components/molecules/Pagination/types.ts
@@ -1,0 +1,8 @@
+import { ButtonProps } from '@/common/components/atoms/Button/Button';
+import { ComponentProps } from 'react';
+import { Link } from 'react-router-dom';
+
+export type PaginationLinkProps = {
+  isActive?: boolean;
+} & Pick<ButtonProps, 'size'> &
+  ComponentProps<typeof Link>;