+++ apps/backoffice-v2/src/common/hooks/useSearch/useSearch.tsx
@@ -1,28 +1,17 @@
-import { AnyArray, TKeyofArrayElement } from '../../types';
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
 import { useDebounce } from '../useDebounce/useDebounce';
-import { search as onSearch } from './search';
-import { useSearchParamsByEntity } from '../useSearchParamsByEntity/useSearchParamsByEntity';
+import { useSerializedSearchParams } from '@/common/hooks/useSerializedSearchParams/useSerializedSearchParams';
 
-export const useSearch = <TArray extends AnyArray>({
-  data,
-  searchBy,
-  initialSearch = '',
-}: {
-  data: TArray;
-  searchBy: Array<TKeyofArrayElement<TArray>>;
-  initialSearch?: string;
-}) => {
-  const [{ search = initialSearch }, setSearchParams] = useSearchParamsByEntity();
-  const searched = useMemo(
-    () =>
-      onSearch({
-        data,
-        searchBy,
-        search,
-      }),
-    [data, searchBy, search],
-  );
+export const useSearch = (
+  {
+    initialSearch = '',
+  }: {
+    initialSearch?: string;
+  } = {
+    initialSearch: '',
+  },
+) => {
+  const [{ search = initialSearch }, setSearchParams] = useSerializedSearchParams();
   const [_search, setSearch] = useState(search);
   const debouncedSearch = useDebounce(_search, 240);
   const onSearchChange = useCallback((search: string) => {
@@ -32,14 +21,12 @@ export const useSearch = <TArray extends AnyArray>({
   useEffect(() => {
     setSearchParams({
       search: debouncedSearch,
-      page: 1,
+      page: '1',
     });
   }, [debouncedSearch]);
 
   return {
-    searched,
     search: _search,
-    searchBy,
     onSearch: onSearchChange,
   };
 };