+++ apps/backoffice-v2/src/pages/TransactionMonitoring/TransactionMonitoring.tsx
@@ -1,0 +1,6 @@
+import React, { FunctionComponent } from 'react';
+import { Outlet } from 'react-router-dom';
+
+export const TransactionMonitoring: FunctionComponent = () => {
+  return <Outlet />;
+};