+++ services/workflows-service/src/customer/customer.controller.internal.ts
@@ -7,7 +7,7 @@ import { Customer } from '@prisma/client';
 import type { TProjectIds } from '@/types';
 import { ProjectIds } from '@/common/decorators/project-ids.decorator';
 
-@swagger.ApiTags('internal/customers')
+@swagger.ApiExcludeController()
 @common.Controller('internal/customers')
 export class CustomerControllerInternal {
   constructor(protected readonly service: CustomerService) {}